# docs/swagger.yaml
basePath: /
definitions:
  web.ErrorResponse:
    properties:
      error: {}
    type: object
  web.HealthcheckResponse:
    properties:
      status:
        type: string
      system_info:
        $ref: '#/definitions/web.HealthcheckSystemInfo'
    type: object
  web.HealthcheckSystemInfo:
    properties:
      environment:
        type: string
      version:
        type: string
    type: object
  web.IndikatorSasaranPenetapanResponse:
    properties:
      definisi_operasional:
        example: definisi abc
        type: string
      id:
        example: 1
        type: integer
      indikator:
        example: Indikator kepuasa publik
        type: string
      kode_indikator:
        example: IND-TUJ-01
        type: string
      rumus_perhitungan:
        example: indeks kepuasan masyarakat
        type: string
      sumber_data:
        example: bps
        type: string
      tahun_aktif:
        example: 2025
        type: integer
      targets:
        items:
          $ref: '#/definitions/web.TargetIndikatorResponse'
        type: array
    type: object
  web.IndikatorTujuanPenetapanResponse:
    properties:
      definisi_operasional:
        example: definisi abc
        type: string
      id:
        example: 1
        type: integer
      indikator:
        example: Indikator kepuasa publik
        type: string
      kode_indikator:
        example: IND-TUJ-01
        type: string
      rumus_perhitungan:
        example: indeks kepuasan masyarakat
        type: string
      sumber_data:
        example: bps
        type: string
      tahun_aktif:
        example: 2025
        type: integer
      targets:
        items:
          $ref: '#/definitions/web.TargetIndikatorResponse'
        type: array
    type: object
  web.RenjaBidangUrusanResponse:
    properties:
      bidang_urusan:
        example: URUSAN PEMERINTAHAN BIDANG PENDIDIKAN
        type: string
      id:
        type: integer
      is_locked:
        example: true
        type: boolean
      kode_bidang_urusan:
        example: "1.01"
        type: string
    type: object
  web.RenjaKegiatanResponse:
    properties:
      id:
        type: integer
      is_locked:
        example: true
        type: boolean
      kegiatan:
        example: KEGIATAN X
        type: string
      kode_kegiatan:
        example: 1.01.01.01.01.01
        type: string
    type: object
  web.RenjaPenetapanOpdResponse:
    properties:
      bidang_urusans:
        items:
          $ref: '#/definitions/web.RenjaBidangUrusanResponse'
        type: array
      is_locked:
        example: true
        type: boolean
      kegiatans:
        items:
          $ref: '#/definitions/web.RenjaKegiatanResponse'
        type: array
      kode_opd:
        example: 1.02.0.00.0.00.01.0000
        type: string
      programs:
        items:
          $ref: '#/definitions/web.RenjaProgramResponse'
        type: array
      subkegiatans:
        items:
          $ref: '#/definitions/web.RenjaSubkegiatanResponse'
        type: array
      tahun_aktif:
        example: 2025
        type: integer
      urusans:
        items:
          $ref: '#/definitions/web.RenjaUrusanResponse'
        type: array
      versi:
        example: 1
        type: integer
    type: object
  web.RenjaProgramResponse:
    properties:
      id:
        type: integer
      is_locked:
        example: true
        type: boolean
      kode_program:
        example: 1.01.01.01
        type: string
      program:
        example: PROGRAM X
        type: string
    type: object
  web.RenjaSubkegiatanResponse:
    properties:
      id:
        type: integer
      is_locked:
        example: true
        type: boolean
      kode_subkegiatan:
        example: 1.01.01.01.01.01.0001
        type: string
      subkegiatan:
        example: SUBKEGIATAN X
        type: string
    type: object
  web.RenjaUrusanResponse:
    properties:
      id:
        type: integer
      is_locked:
        example: true
        type: boolean
      kode_urusan:
        example: "1"
        type: string
      urusan:
        example: URUSAN PENUNJANG
        type: string
    type: object
  web.Response-web_HealthcheckResponse:
    properties:
      data:
        $ref: '#/definitions/web.HealthcheckResponse'
    type: object
  web.Response-web_RenjaPenetapanOpdResponse:
    properties:
      data:
        $ref: '#/definitions/web.RenjaPenetapanOpdResponse'
    type: object
  web.Response-web_SasaranPenetapanOpdResponse:
    properties:
      data:
        $ref: '#/definitions/web.SasaranPenetapanOpdResponse'
    type: object
  web.Response-web_SyncPenetapanOpdResponse:
    properties:
      data:
        $ref: '#/definitions/web.SyncPenetapanOpdResponse'
    type: object
  web.Response-web_TujuanPenetapanOpdResponse:
    properties:
      data:
        $ref: '#/definitions/web.TujuanPenetapanOpdResponse'
    type: object
  web.SasaranOpdResponse:
    properties:
      id:
        type: integer
      indikators:
        items:
          $ref: '#/definitions/web.IndikatorSasaranPenetapanResponse'
        type: array
      kode_sasaran_opd:
        example: SAS-001
        type: string
      periode:
        example: 2025-2029
        type: string
      sasaran_opd:
        example: Meningkatnya Nilai SAKIP
        type: string
    type: object
  web.SasaranPenetapanOpdResponse:
    properties:
      is_locked:
        example: true
        type: boolean
      kode_opd:
        example: 1.02.0.00.0.00.01.0000
        type: string
      sasaran_opds:
        items:
          $ref: '#/definitions/web.SasaranOpdResponse'
        type: array
      tahun_aktif:
        example: 2025
        type: integer
      versi:
        example: 1
        type: integer
    type: object
  web.SyncPenetapanOpdRequest:
    properties:
      kode_opd:
        example: 1.22.33
        type: string
      tahun:
        example: 2025
        type: integer
    type: object
  web.SyncPenetapanOpdResponse:
    properties:
      jenis_penetapan:
        example: TUJUAN-OPD
        type: string
      kode_opd:
        example: 1.22.33
        type: string
      processed_at:
        type: string
      processed_summary:
        $ref: '#/definitions/web.SyncPenetapanOpdSummary'
      status:
        example: SUCCESS
        type: string
      sync_id:
        example: 155
        type: integer
      tahun:
        example: 2025
        type: integer
    type: object
  web.SyncPenetapanOpdSummary:
    properties:
      indikator:
        example: 3
        type: integer
      renja:
        example: 1
        type: integer
      sasaran:
        example: 1
        type: integer
      target:
        example: 4
        type: integer
      tujuan:
        example: 1
        type: integer
    type: object
  web.TargetIndikatorResponse:
    properties:
      id:
        example: 1
        type: integer
      kode_target:
        example: TGT-IND-01
        type: string
      satuan:
        example: '%'
        type: string
      tahun:
        example: 2025
        type: integer
      target:
        example: 100
        type: number
    type: object
  web.TujuanOpdResponse:
    properties:
      id:
        example: 1
        type: integer
      indikators:
        items:
          $ref: '#/definitions/web.IndikatorTujuanPenetapanResponse'
        type: array
      kode_tujuan_opd:
        example: TUJ-001
        type: string
      periode:
        example: 2025-2029
        type: string
      tujuan_opd:
        example: Meningkatkan kualitas pelayanan publik
        type: string
    type: object
  web.TujuanPenetapanOpdResponse:
    properties:
      is_locked:
        example: true
        type: boolean
      kode_opd:
        example: 1.02.0.00.0.00.01.0000
        type: string
      tahun_aktif:
        example: 2025
        type: integer
      tujuan_opds:
        items:
          $ref: '#/definitions/web.TujuanOpdResponse'
        type: array
      versi:
        example: 1
        type: integer
    type: object
  web.ValidationErrorResponse:
    properties:
      error:
        additionalProperties:
          type: string
        example:
          kode_opd: required
          tahun: required
        type: object
    type: object
info:
  contact:
    name: Kertaskerja Dev Team
  description: Service snapshot penetapan tujuan, sasaran, renja, rekin untuk pemda,
    opd, dan individu
  termsOfService: http://swagger.io/terms/
  title: Penetapan Service
  version: "1.0"
paths:
  /healthcheck:
    get:
      description: checking if this service is accessible
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/web.Response-web_HealthcheckResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/web.ErrorResponse'
      summary: healthcheck
      tags:
      - healthcheck
  /opd/renja:
    get:
      consumes:
      - application/json
      description: Mengambil data renja OPD berdasarkan kode OPD dan tahun penetapan
      parameters:
      - description: Kode OPD
        in: query
        name: kodeOpd
        required: true
        type: string
      - description: Tahun Penetapan
        in: query
        name: tahun
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: Berhasil mengambil data renja OPD
          schema:
            $ref: '#/definitions/web.Response-web_RenjaPenetapanOpdResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/web.ValidationErrorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/web.ErrorResponse'
      summary: Get sasaran OPD penetapan
      tags:
      - OPD
  /opd/renja/sync:
    post:
      consumes:
      - application/json
      description: Sinkron data penetapan renja OPD berdasarkan kode OPD dan tahun
        dari perencanaan
      parameters:
      - description: Payload sinkronisasi penetapan OPD
        in: body
        name: payload
        required: true
        schema:
          $ref: '#/definitions/web.SyncPenetapanOpdRequest'
      produces:
      - application/json
      responses:
        "200":
          description: Success
          schema:
            $ref: '#/definitions/web.Response-web_SyncPenetapanOpdResponse'
        "422":
          description: Unprocessable Entity
          schema:
            $ref: '#/definitions/web.ValidationErrorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/web.ErrorResponse'
      summary: Sync Penetapan Renja OPD
      tags:
      - Sync
  /opd/sasaran:
    get:
      consumes:
      - application/json
      description: Mengambil data sasaran OPD berdasarkan kode OPD dan tahun penetapan
      parameters:
      - description: Kode OPD
        in: query
        name: kodeOpd
        required: true
        type: string
      - description: Tahun Penetapan
        in: query
        name: tahun
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: Berhasil mengambil data sasaran OPD
          schema:
            $ref: '#/definitions/web.Response-web_SasaranPenetapanOpdResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/web.ValidationErrorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/web.ErrorResponse'
      summary: Get sasaran OPD penetapan
      tags:
      - OPD
  /opd/sasaran/sync:
    post:
      consumes:
      - application/json
      description: Sinkron data penetapan sasaran OPD berdasarkan kode OPD dan tahun
        dari perencanaan
      parameters:
      - description: Payload sinkronisasi penetapan OPD
        in: body
        name: payload
        required: true
        schema:
          $ref: '#/definitions/web.SyncPenetapanOpdRequest'
      produces:
      - application/json
      responses:
        "200":
          description: Success
          schema:
            $ref: '#/definitions/web.Response-web_SyncPenetapanOpdResponse'
        "422":
          description: Unprocessable Entity
          schema:
            $ref: '#/definitions/web.ValidationErrorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/web.ErrorResponse'
      summary: Sync Penetapan Sasaran OPD
      tags:
      - Sync
  /opd/tujuan:
    get:
      consumes:
      - application/json
      description: Mengambil data tujuan OPD berdasarkan kode OPD dan tahun penetapan
      parameters:
      - description: Kode OPD
        in: query
        name: kodeOpd
        required: true
        type: string
      - description: Tahun Penetapan
        in: query
        name: tahun
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: Berhasil mengambil data tujuan OPD
          schema:
            $ref: '#/definitions/web.Response-web_TujuanPenetapanOpdResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/web.ValidationErrorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/web.ErrorResponse'
      summary: Get tujuan OPD penetapan
      tags:
      - OPD
  /opd/tujuan/sync:
    post:
      consumes:
      - application/json
      description: Sinkron data penetapan tujuan OPD berdasarkan kode OPD dan tahun
        dari perencanaan
      parameters:
      - description: Payload sinkronisasi penetapan OPD
        in: body
        name: payload
        required: true
        schema:
          $ref: '#/definitions/web.SyncPenetapanOpdRequest'
      produces:
      - application/json
      responses:
        "200":
          description: Success
          schema:
            $ref: '#/definitions/web.Response-web_SyncPenetapanOpdResponse'
        "422":
          description: Unprocessable Entity
          schema:
            $ref: '#/definitions/web.ValidationErrorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/web.ErrorResponse'
      summary: Sync Penetapan Tujuan OPD
      tags:
      - Sync
schemes:
- http
- https
swagger: "2.0"